sc-adapter/requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.14.0
langdetect==1.0.9
py-sc-client==0.2.4
py-sc-kpm==0.0.1
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
websocket-client==1.4.2
